devtools/server/tests/mochitest/test_inspector-changeattrs.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
};

let gInspectee = null;
let gWalker = null;

addTest(async function setup() {
  const url = document.getElementById("inspectorContent").href;
  const { target, doc } = await attachURL(url);
  gInspectee = doc;
  const inspector = await target.getInspector();
  gWalker = inspector.walker;
  runNextTest();
});

addTest(function testChangeAttrs() {
  const attrNode = gInspectee.querySelector("#a");
  let attrFront;
  promiseDone(gWalker.querySelector(gWalker.rootNode, "#a").then(front => {
    attrFront = front;
    dump("attrFront is: " + attrFront + "\n");
    // Add a few attributes.
    const list = attrFront.startModifyingAttributes();
    list.setAttribute("data-newattr", "newvalue");
    list.setAttribute("data-newattr2", "newvalue");
    return list.apply();
  }).then(() => {
    // We're only going to test that the change hit the document.
    // There are other tests that make sure changes are propagated
    // to the client.
    is(attrNode.getAttribute("data-newattr"), "newvalue",
       "Node should have the first new attribute");
    is(attrNode.getAttribute("data-newattr2"), "newvalue",
       "Node should have the second new attribute.");
  }).then(() => {
    // Change an attribute.
    const list = attrFront.startModifyingAttributes();
    list.setAttribute("data-newattr", "changedvalue");
    return list.apply();
  }).then(() => {
    is(attrNode.getAttribute("data-newattr"), "changedvalue",
       "Node should have the changed first value.");
    is(attrNode.getAttribute("data-newattr2"), "newvalue",
       "Second value should remain unchanged.");
  }).then(() => {
    const list = attrFront.startModifyingAttributes();
    list.removeAttribute("data-newattr2");
    return list.apply();
  }).then(() => {
    is(attrNode.getAttribute("data-newattr"), "changedvalue",
       "Node should have the changed first value.");
    ok(!attrNode.hasAttribute("data-newattr2"), "Second value should be removed.");
  }).then(runNextTest));
});

addTest(function cleanup() {
  gWalker = null;
  gInspectee = null;
  runNextTest();
});
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
